+++ ui/api/topics/actions.ts
@@ -1,146 +1,24 @@
-import { BackendError, getHeaders } from "@/api/api";
+"use server";
+import { getHeaders } from "@/api/api";
+import { getKafkaCluster } from "@/api/kafka";
+import {
+  describeTopicsQuery,
+  NewConfigMap,
+  Topic,
+  TopicCreateResponse,
+  TopicCreateResponseSchema,
+  TopicMutateError,
+  TopicMutateResponseErrorSchema,
+  TopicResponse,
+  TopicsResponse,
+  TopicsResponseList,
+} from "@/api/topics/schema";
 import { filterUndefinedFromObj } from "@/utils/filterUndefinedFromObj";
 import { logger } from "@/utils/logger";
+import { getSession, setSession } from "@/utils/session";
 import { revalidateTag } from "next/cache";
-import { z } from "zod";
 
 const log = logger.child({ module: "topics-api" });
-
-const describeTopicsQuery = encodeURI(
-  "fields[topics]=,name,internal,partitions,authorizedOperations,configs,recordCount,totalLeaderLogBytes",
-);
-
-const OffsetSchema = z.object({
-  offset: z.number().optional(),
-  timestamp: z.string().optional(),
-  leaderEpoch: z.number().optional(),
-});
-const PartitionSchema = z.object({
-  partition: z.number(),
-  leaderId: z.number(),
-  replicas: z.array(
-    z.object({
-      nodeId: z.number(),
-      nodeRack: z.string().optional(),
-      inSync: z.boolean(),
-      localStorage: BackendError.or(
-        z.object({
-          size: z.number(),
-          offsetLag: z.number(),
-          future: z.boolean(),
-        }),
-      ).optional(),
-    }),
-  ),
-  offsets: z
-    .object({
-      earliest: OffsetSchema.optional(),
-      latest: OffsetSchema.optional(),
-      maxTimestamp: OffsetSchema.optional(),
-      timestamp: OffsetSchema.optional(),
-    })
-    .optional()
-    .nullable(),
-  recordCount: z.number().optional(),
-  leaderLocalStorage: z.number().optional(),
-});
-const ConfigSchema = z.object({
-  value: z.string(),
-  source: z.string(),
-  sensitive: z.boolean(),
-  readOnly: z.boolean(),
-  type: z.string(),
-});
-
-const ConfigMapSchema = z.record(z.string(), ConfigSchema);
-export type ConfigMap = z.infer<typeof ConfigMapSchema>;
-const NewConfigMapSchema = z.record(
-  z.string(),
-  z.object({
-    value: z.union([z.string(), z.number(), z.undefined(), z.null()]),
-  }),
-);
-export type NewConfigMap = z.infer<typeof NewConfigMapSchema>;
-
-const TopicSchema = z.object({
-  id: z.string(),
-  type: z.literal("topics"),
-  attributes: z.object({
-    name: z.string(),
-    internal: z.boolean(),
-    partitions: z.array(PartitionSchema),
-    authorizedOperations: z.array(z.string()),
-    configs: ConfigMapSchema,
-    recordCount: z.number().optional(),
-    totalLeaderLogBytes: z.number().optional(),
-  }),
-});
-export const TopicResponse = z.object({
-  data: TopicSchema,
-});
-export type Topic = z.infer<typeof TopicSchema>;
-
-const TopicListSchema = z.object({
-  id: z.string(),
-  type: z.literal("topics"),
-  meta: z.object({
-    page: z.object({
-      cursor: z.string(),
-    }),
-  }),
-  attributes: TopicSchema.shape.attributes.pick({
-    name: true,
-    internal: true,
-    partitions: true,
-    recordCount: true,
-    totalLeaderLogBytes: true,
-  }),
-});
-export type TopicList = z.infer<typeof TopicListSchema>;
-export const TopicsResponse = z.object({
-  meta: z.object({
-    page: z.object({
-      total: z.number(),
-      pageNumber: z.number(),
-    }),
-  }),
-  links: z.object({
-    first: z.string().nullable(),
-    prev: z.string().nullable(),
-    next: z.string().nullable(),
-    last: z.string().nullable(),
-  }),
-  data: z.array(TopicListSchema),
-});
-export type TopicsResponseList = z.infer<typeof TopicsResponse>;
-
-const TopicCreateResponseSuccessSchema = z.object({
-  data: z.object({
-    id: z.string(),
-  }),
-});
-const TopicMutateResponseErrorSchema = z.object({
-  errors: z.array(
-    z.object({
-      id: z.string(),
-      status: z.string(),
-      code: z.string(),
-      title: z.string(),
-      detail: z.string(),
-      source: z
-        .object({
-          pointer: z.string().optional(),
-        })
-        .optional(),
-    }),
-  ),
-});
-const TopicCreateResponseSchema = z.union([
-  TopicCreateResponseSuccessSchema,
-  TopicMutateResponseErrorSchema,
-]);
-export type TopicMutateError = z.infer<typeof TopicMutateResponseErrorSchema>;
-export type TopicCreateResponse = z.infer<typeof TopicCreateResponseSchema>;
 
 export async function getTopics(
   kafkaId: string,
@@ -291,4 +169,55 @@ export async function deleteTopic(
     log.error(e, "deleteTopic unknown error");
   }
   return false;
+}
+
+type ViewedTopicsSession = { viewedTopics: ViewedTopic[] | undefined };
+export type ViewedTopic = {
+  kafkaId: string;
+  kafkaName: string;
+  topicId: string;
+  topicName: string;
+};
+
+export async function getViewedTopics(): Promise<ViewedTopic[]> {
+  log.info("getViewedTopics");
+  const recentTopicsSession =
+    await getSession<ViewedTopicsSession>("recent-topics");
+  log.debug(recentTopicsSession, "getViewedTopics session");
+  return recentTopicsSession.viewedTopics || [];
+}
+
+export async function setTopicAsViewed(kafkaId: string, topicId: string) {
+  log.info({ kafkaId, topicId }, "setTopicAsViewed");
+  const cluster = await getKafkaCluster(kafkaId);
+  const topic = await getTopic(kafkaId, topicId);
+  const viewedTopics = await getViewedTopics();
+  if (cluster && topic) {
+    const viewedTopic: ViewedTopic = {
+      kafkaId,
+      kafkaName: cluster.attributes.name,
+      topicId,
+      topicName: topic.attributes.name,
+    };
+    if (viewedTopics.find((t) => t.topicId === viewedTopic.topicId)) {
+      log.debug(
+        { kafkaId, topicId },
+        "setTopicAsViewed: topic was already in the list, ignoring",
+      );
+      return viewedTopics;
+    }
+    log.debug(
+      { kafkaId, topicId },
+      "setTopicAsViewed: adding topic to the list",
+    );
+    const updatedViewedTopics = [viewedTopic, ...viewedTopics].slice(0, 5);
+    await setSession<ViewedTopicsSession>("recent-topics", {
+      viewedTopics: updatedViewedTopics,
+    });
+    log.debug(updatedViewedTopics, "setTopicAsViewed: updated list");
+    return updatedViewedTopics;
+  } else {
+    log.debug({ topic, cluster }, "setTopicAsViewed: invalid topic/cluster");
+    return viewedTopics;
+  }
 }
