+++ ui/app/[locale]/kafka/[kafkaId]/@activeBreadcrumb/topics/[topicId]/TopicBreadcrumb.tsx
@@ -1,4 +1,4 @@
-import { getTopic } from "@/api/topics";
+import { getTopic } from "@/api/topics/actions";
 import { KafkaTopicParams } from "@/app/[locale]/kafka/[kafkaId]/topics/kafkaTopic.params";
 import { BreadcrumbLink } from "@/components/BreadcrumbLink";
 import { BreadcrumbItem } from "@/libs/patternfly/react-core";
