+++ ui/app/[locale]/kafka/[kafkaId]/@header/KafkaHeader.tsx
@@ -1,5 +1,5 @@
 import { getKafkaCluster } from "@/api/kafka";
-import { getTopics } from "@/api/topics";
+import { getTopics } from "@/api/topics/actions";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
 import { AppHeader } from "@/components/AppHeader";
 import { Skeleton } from "@patternfly/react-core";
