+++ ui/app/[locale]/kafka/[kafkaId]/page.tsx
@@ -1,12 +1,6 @@
-import { getKafkaCluster } from "@/api/kafka";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
 import { redirect } from "@/navigation";
-import { notFound } from "next/navigation";
 
-export default async function KafkaRoot({ params }: { params: KafkaParams }) {
-  const cluster = await getKafkaCluster(params.kafkaId);
-  if (!cluster) {
-    notFound();
-  }
-  redirect(`${params.kafkaId}/topics`);
+export default function KafkaRoot({ params }: { params: KafkaParams }) {
+  redirect(`/kafka/${params.kafkaId}/topics`);
 }
