+++ ui/app/[locale]/kafka/[kafkaId]/topics/[topicId]/configuration/page.tsx
@@ -1,4 +1,4 @@
-import { getTopic, updateTopic } from "@/api/topics";
+import { getTopic, updateTopic } from "@/api/topics/actions";
 import { KafkaTopicParams } from "@/app/[locale]/kafka/[kafkaId]/topics/kafkaTopic.params";
 import { PageSection } from "@/libs/patternfly/react-core";
 import { Suspense } from "react";
