+++ ui/app/[locale]/kafka/[kafkaId]/topics/[topicId]/partitions/page.tsx
@@ -1,4 +1,4 @@
-import { getTopic } from "@/api/topics";
+import { getTopic } from "@/api/topics/actions";
 import { KafkaTopicParams } from "@/app/[locale]/kafka/[kafkaId]/topics/kafkaTopic.params";
 import { notFound } from "next/navigation";
 import { Suspense } from "react";
