+++ ui/app/[locale]/kafka/[kafkaId]/topics/create/StepOptions.tsx
@@ -1,4 +1,4 @@
-import { ConfigMap, NewConfigMap, TopicMutateError } from "@/api/topics";
+import { ConfigMap, NewConfigMap, TopicMutateError } from "@/api/topics/schema";
 import { ConfigTable } from "@/app/[locale]/kafka/[kafkaId]/topics/create/ConfigTable";
 import { Error } from "@/app/[locale]/kafka/[kafkaId]/topics/create/Errors";
 import { topicMutateErrorToFieldError } from "@/app/[locale]/kafka/[kafkaId]/topics/create/topicMutateErrorToFieldError";
