+++ ui/app/[locale]/kafka/[kafkaId]/topics/create/StepReview.tsx
@@ -1,4 +1,4 @@
-import { ConfigMap, NewConfigMap, TopicMutateError } from "@/api/topics";
+import { ConfigMap, NewConfigMap, TopicMutateError } from "@/api/topics/schema";
 import { Error } from "@/app/[locale]/kafka/[kafkaId]/topics/create/Errors";
 import { ReviewTable } from "@/app/[locale]/kafka/[kafkaId]/topics/create/ReviewTable";
 import { Number } from "@/components/Number";
