+++ ui/app/[locale]/kafka/[kafkaId]/topics/create/page.tsx
@@ -1,5 +1,6 @@
 import { getKafkaCluster } from "@/api/kafka";
-import { createTopic, NewConfigMap } from "@/api/topics";
+import { createTopic } from "@/api/topics/actions";
+import { NewConfigMap } from "@/api/topics/schema";
 import { KafkaParams } from "@/app/[locale]/kafka/[kafkaId]/kafka.params";
 import { CreateTopic } from "@/app/[locale]/kafka/[kafkaId]/topics/create/CreateTopic";
 import { redirect } from "@/navigation";
