--- ui/app/[locale]/resources/create/kafka/authentication/page.tsx
@@ -1,13 +1,0 @@
-import { getSession } from "@/utils/session";
-import { redirect } from "@/navigation";
-import { Step2 } from "./Step2";
-
-export default async function AsyncCreateResourceStep2Page() {
-  const session = await getSession("resources");
-  const newResource = session?.newResource;
-  const { name, boostrapServer } = newResource || {};
-  if (!name || !boostrapServer) {
-    redirect("/");
-  }
-  return <Step2 />;
-}
